Finished Constructors & getShapeType() Methods
The empty constructors have been filled along with the getShapeType() Methods for the Circle, Ellipse, EquilateralTriangle, Polygon, and Trapezoid classes. 

I have not filled out the Square and Rectangle classes because those classes are initially empty while the classes listed above required me to simply fill in what was missing. This step will be done in the near future.