Square & Rectangle Classes Rough Drafts
Rough drafts of the square and rectangle class. These classes and their methods were made following the UML Diagram in the Lab 6 documentation. These are the only two classes that we were required to make from scratch in this program. There will likely be later edits to correct small errors.